+++ Code/Engine/Core/Scripting/ScriptWorldModule.h
@@ -1,5 +1,6 @@
 #pragma once
 
+#include <Core/ResourceManager/ResourceHandle.h>
 #include <Core/Utils/IntervalScheduler.h>
 #include <Core/World/World.h>
 
@@ -45,4 +46,19 @@ private:
   ezHashTable<ezScriptClassResourceHandle, ReloadFunctionList> m_ReloadFunctions;
   ezHashSet<ezScriptClassResourceHandle> m_NeedReload;
   ReloadFunctionList m_TempReloadFunctions;
+};
+
+//////////////////////////////////////////////////////////////////////////
+
+template <>
+struct ezHashHelper<ezScriptWorldModule::FunctionContext>
+{
+  EZ_ALWAYS_INLINE static ezUInt32 Hash(const ezScriptWorldModule::FunctionContext& value)
+  {
+    ezUInt32 hash = ezHashHelper<const void*>::Hash(value.m_pFunction);
+    hash = ezHashingUtils::CombineHashValues32(hash, ezHashHelper<void*>::Hash(value.m_pInstance));
+    return hash;
+  }
+
+  EZ_ALWAYS_INLINE static bool Equal(const ezScriptWorldModule::FunctionContext& a, const ezScriptWorldModule::FunctionContext& b) { return a == b; }
 };
